requirements.txt
cx-Oracle==8.2.1
Django==3.2.7
django-crispy-forms
menuinst==1.4.16
Pillow==8.3.2
django-filter==21.1
